% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invariance_test.R
\name{invariance_test}
\alias{invariance_test}
\alias{invariance_test.default}
\alias{invariance_test.tbl_df}
\alias{invariance_test.xts}
\alias{invariance_test.matrix}
\alias{invariance_test.numeric}
\title{Visual Test for Stationarity}
\usage{
invariance_test(.invariant)

\method{invariance_test}{default}(.invariant)

\method{invariance_test}{tbl_df}(.invariant)

\method{invariance_test}{xts}(.invariant)

\method{invariance_test}{matrix}(.invariant)

\method{invariance_test}{numeric}(.invariant)
}
\arguments{
\item{.invariant}{An univariate time-series.}
}
\value{
A \code{ggplot2} object.
}
\description{
For "invariant" time-series the histograms should look the same and the
shape of the ellipsoid should be circular.
}
\examples{
library(tibble)

x <- as_tibble(diff(log(EuStockMarkets)))

invariance_test(x["CAC"])
invariance_test(x["DAX"])
}
