% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{ks_test}
\alias{ks_test}
\alias{ks_test.default}
\alias{ks_test.tbl}
\alias{ks_test.xts}
\alias{ks_test.matrix}
\title{Kolmogorov-Smirnov Test for Invariance}
\usage{
ks_test(.invariant, ...)

\method{ks_test}{default}(.invariant, ...)

\method{ks_test}{tbl}(.invariant, ...)

\method{ks_test}{xts}(.invariant, ...)

\method{ks_test}{matrix}(.invariant, ...)
}
\arguments{
\item{.invariant}{An univariate timeseries.}

\item{...}{Additional arguments to be passed to \code{\link[stats]{ks.test}}.}
}
\value{
A ggplot2 object.
}
\description{
Perform two-sample Kolmogorov-Smirnov test.
}
\details{
This function wrappers \code{\link[stats]{ks.test}}.
}
\examples{
x <- diff(log(EuStockMarkets))[ , 2]

ks_test(matrix(x))
}
