% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid.R
\name{plot_ellipsoid}
\alias{plot_ellipsoid}
\title{Two Dimensional Ellipsoid}
\usage{
plot_ellipsoid(.invariant)
}
\arguments{
\item{.invariant}{An univariate \code{tibble}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Visualize the two dimensional ellipsoid for invariance test.
}
\examples{
#
}
