% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stats.R
\name{series_describe}
\alias{series_describe}
\title{Summary Statistics of Time-Series}
\usage{
series_describe(.invariant, .digits = 4)
}
\arguments{
\item{.invariant}{An univariate or multivariate time-series.}

\item{.digits}{The number of digits to print.}
}
\value{
A \code{tibble} with basic statistics for exploratory analysis.
}
\description{
This function is designed for exploratory purposes.
}
\examples{
data(edhec)

# univariate
series_describe(diff(log(EuStockMarkets[ , 1, drop = FALSE])))

# multivariate
series_describe(diff(log(EuStockMarkets)))
}
\seealso{
\code{\link{skewness}}, \code{\link{kurtosis}}
}
